Refactored stuff
