+++ core/src/com/github/va7icana55asin/test_game/screen/Main.java
@@ -1,8 +1,9 @@
-package com.github.va7icana55asin.test_game;
+package com.github.va7icana55asin.test_game.screen;
 
 import com.badlogic.gdx.*;
 import com.badlogic.gdx.graphics.g2d.BitmapFont;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
+import com.github.va7icana55asin.test_game.screen.MainMenuScreen;
 
 public class Main extends Game {
 	protected SpriteBatch batch;
