+++ core/src/com/github/va7icana55asin/test_game/screen/MainMenuScreen.java
@@ -1,38 +1,46 @@
-package com.github.va7icana55asin.test_game;
+package com.github.va7icana55asin.test_game.screen;
 
 import com.badlogic.gdx.Gdx;
+import com.badlogic.gdx.Input;
 import com.badlogic.gdx.Screen;
 import com.badlogic.gdx.graphics.GL20;
 import com.badlogic.gdx.graphics.OrthographicCamera;
+import com.github.va7icana55asin.test_game.utility.DIFFICULTY;
 
-public class GameOverScreen implements Screen {
+public class MainMenuScreen implements Screen {
     private final Main game;
 
     private OrthographicCamera camera;
 
-    public GameOverScreen(final Main game){
+    public MainMenuScreen(final Main game) {
         this.game = game;
 
         camera = new OrthographicCamera();
-        camera.setToOrtho(false, 640f,480f);
+        camera.setToOrtho(false,640f,480f);
     }
 
     @Override
-    public void show() {
-
-    }
-
-    @Override
-    public void render(float delta) {
-        Gdx.gl.glClearColor(1, 0, 0, 1);
+    public void render(float delta){
+        Gdx.gl.glClearColor(0, 0, 0.2f, 1);
         Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT);
 
         camera.update();
         game.batch.setProjectionMatrix(camera.combined);
 
         game.batch.begin();
-        game.font.draw(game.batch, "You Lost", 100, 150);
+        game.font.draw(game.batch, "Welcome to Test Game!!! ", 100, 150);
+        game.font.draw(game.batch, "Tap anywhere or press \'enter\' to begin!", 100, 100);
         game.batch.end();
+
+        if (Gdx.input.isTouched() || Gdx.input.isKeyPressed(Input.Keys.ENTER)) {
+            game.setScreen(new GameScreen(game, DIFFICULTY.EASY));
+            dispose();
+        }
+    }
+
+    @Override
+    public void show() {
+
     }
 
     @Override
