+++ core/src/com/github/va7icana55asin/test_game/utility/DIFFICULTY.java
@@ -1,4 +1,4 @@
-package com.github.va7icana55asin.test_game;
+package com.github.va7icana55asin.test_game.utility;
 
 public enum DIFFICULTY {
     BABY,EASY,MEDIUM,HARD,EXTREME
