+++ desktop/src/com/github/va7icana55asin/test_game/desktop/DesktopLauncher.java
@@ -2,7 +2,7 @@ package com.github.va7icana55asin.test_game.desktop;
 
 import com.badlogic.gdx.backends.lwjgl.LwjglApplication;
 import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration;
-import com.github.va7icana55asin.test_game.Main;
+import com.github.va7icana55asin.test_game.screen.Main;
 
 public class DesktopLauncher {
 	public static void main (String[] arg) {
